Bug 838657 - Implement a console reporter. r=julienw,eli
Rebased from https://github.com/mozilla-b2g/gaia/pull/11870 by @caiolima
