+++ Makefile
@@ -484,19 +484,19 @@ define app-makefile-template
 .PHONY: $(1)
 $(1): $(XULRUNNER_BASE_DIRECTORY) keyboard-layouts contacts-import-services $(STAGE_DIR)/settings_stage.json webapp-manifests svoperapps clear-stage-app webapp-shared | $(STAGE_DIR)
 	@if [[ ("$(2)" =~ "${BUILD_APP_NAME}") || ("${BUILD_APP_NAME}" == "*") ]]; then \
-  	if [ -r "$(2)/Makefile" ]; then \
-  		echo "execute Makefile for $(1) app" ; \
-  		STAGE_APP_DIR="../../build_stage/$(1)" make -C "$(2)" ; \
-  	else \
-  		echo "copy $(1) to build_stage/" ; \
-  		cp -LR "$(2)" $(STAGE_DIR) && \
-  		if [ -r "$(2)/build/build.js" ]; then \
-  			echo "execute $(1)/build/build.js"; \
-  			export APP_DIR=$(2); \
-  			$(call run-js-command,app/build); \
-  		fi; \
-  	fi && \
-  	$(call clean-build-files,$(STAGE_DIR)/$(1)); \
+	if [ -r "$(2)/Makefile" ]; then \
+		echo "execute Makefile for $(1) app" ; \
+		STAGE_APP_DIR="../../build_stage/$(1)" make -C "$(2)" ; \
+	else \
+		echo "copy $(1) to build_stage/" ; \
+		cp -LR "$(2)" $(STAGE_DIR) && \
+		if [ -r "$(2)/build/build.js" ]; then \
+			echo "execute $(1)/build/build.js"; \
+			export APP_DIR=$(2); \
+			$(call run-js-command,app/build); \
+		fi; \
+	fi && \
+	$(call clean-build-files,$(STAGE_DIR)/$(1)); \
   fi;
 endef
 
@@ -818,7 +818,10 @@ caldav-server-install:
 
 .PHONY: test-perf
 test-perf:
-	MOZPERFOUT="$(MOZPERFOUT)" APPS="$(APPS)" MARIONETTE_RUNNER_HOST=$(MARIONETTE_RUNNER_HOST) GAIA_DIR="`pwd`" ./bin/gaia-perf-marionette
+	MOZPERFOUT="$(MOZPERFOUT)" APPS="$(APPS)" \
+	MARIONETTE_RUNNER_HOST=$(MARIONETTE_RUNNER_HOST) GAIA_DIR="`pwd`" \
+	REPORTER=$(REPORTER) \
+	./bin/gaia-perf-marionette
 
 .PHONY: tests
 tests: app-makefiles offline
